backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
pandas>=2.2.0
openpyxl>=3.1.2
google-genai
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
plotly>=5.18.0
numpy>=1.26.0
httpx>=0.26.0
openai==2.15.0
